pyproject.toml
[tool.poetry]
name = "abctk"
version = "0.1.1"
description = ""
authors = ["Nori Hayashi <net@hayashi-lin.net>"]
maintainers = ["Nori Hayashi <net@hayashi-lin.net>"]
readme = "README.md"
repository = "https://github.com/ABCTreebank/ABCTreebank"
packages = [
    { include = "abctk" },
    { include = "lit", format = "sdist" }
]

include = ["ext_src/", "ext_scripts/"]
exclude = ["**/docs"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/ABCTreebank/ABCTreebank/issues"

[tool.poetry.scripts]
abctk = "abctk.cli:cmd_main" 

[tool.poetry.dependencies]
python = "~3.6.1"
click = "^7.1.2"
xdg = "^5.0.1"
parsy = "^1.3.0"
numpy = "^1.19.4"
Janome = "^0.4.1"
"ruamel.yaml" = "^0.16.12"
dictknife = "^0.13.0"
psutil = "^5.8.0"
tqdm = "^4.54.1"
allennlp = { version = "^0.9.0", optional = true } 
depccg = "^1.1.0"
GitPython = "^3.1.11"
gevent = { version = "~1.4.0", optional = true }
importlib-metadata = "^3.3.0"
pycuda = { version = "^2020.1", optional = true }
attrs = "^20.3.0"
lxml = "^4.6.2"
typing-extensions = {version = "^3.7.4", python = ">=3.4 <3.7"}
lark-parser = "^0.11.2"
importlib-resources = {version = "^5.1.2", python = "<3.7"}
fs = "^2.4.12"
humanize = "^3.3.0"
more-itertools = "^8.7.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
requests = "^2.25.1"
mypy = "^0.790"
pytest = "^6.2.2"
Sphinx = "^3.5.3"
sphinx-rtd-theme = "^0.5.1"
ipykernel = "^5.5.0"
pandas = "~1.1"

[tool.poetry.extras]
parser = ["allennlp", "gevent"]
ml = ["allennlp", "gevent", "pycuda"]

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[build-system]
requires = ["poetry-core>=1.0.0", "requests", "xdg", "tqdm"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]